LPS-164095 Rename and use the proper import
