+++ modules/apps/notification/notification-api/src/main/java/com/liferay/notification/service/NotificationTemplateLocalService.java
@@ -15,8 +15,8 @@
 package com.liferay.notification.service;
 
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
+import com.liferay.notification.context.NotificationContext;
 import com.liferay.notification.model.NotificationTemplate;
-import com.liferay.notification.type.NotificationContext;
 import com.liferay.petra.sql.dsl.query.DSLQuery;
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.DynamicQuery;
