+++ modules/apps/notification/notification-api/src/main/java/com/liferay/notification/service/NotificationTemplateServiceUtil.java
@@ -37,7 +37,7 @@ public class NotificationTemplateServiceUtil {
 	 * Never modify this class directly. Add custom service methods to <code>com.liferay.notification.service.impl.NotificationTemplateServiceImpl</code> and rerun ServiceBuilder to regenerate this class.
 	 */
 	public static NotificationTemplate addNotificationTemplate(
-			com.liferay.notification.type.NotificationContext
+			com.liferay.notification.context.NotificationContext
 				notificationContext)
 		throws PortalException {
 
@@ -75,7 +75,7 @@ public class NotificationTemplateServiceUtil {
 	}
 
 	public static NotificationTemplate updateNotificationTemplate(
-			com.liferay.notification.type.NotificationContext
+			com.liferay.notification.context.NotificationContext
 				notificationContext)
 		throws PortalException {
