+++ modules/apps/notification/notification-api/src/main/resources/com/liferay/notification/constants/packageinfo
@@ -1,1 +1,1 @@
-version 2.1.0
+version 2.2.0
