+++ modules/apps/notification/notification-service/src/main/java/com/liferay/notification/internal/type/EmailNotificationType.java
@@ -423,8 +423,8 @@ public class EmailNotificationType extends BaseNotificationType {
 
 		return formatLocalizedContent(
 			StringUtil.merge(emailAddresses), locale,
-			notificationContext,
-			NotificationTermContributorConstants.RECIPIENT);
+			NotificationTermContributorConstants.RECIPIENT,
+			notificationContext);
 	}
 
 	private List<Long> _getFileEntryIds(
