+++ modules/apps/notification/notification-service/src/main/java/com/liferay/notification/internal/type/users/provider/DefaultUsersProvider.java
@@ -15,9 +15,9 @@
 package com.liferay.notification.internal.type.users.provider;
 
 import com.liferay.notification.constants.NotificationRecipientConstants;
+import com.liferay.notification.context.NotificationContext;
 import com.liferay.notification.model.NotificationRecipient;
 import com.liferay.notification.model.NotificationTemplate;
-import com.liferay.notification.type.NotificationContext;
 import com.liferay.petra.function.transform.TransformUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.User;
