+++ modules/apps/notification/notification-service/src/main/java/com/liferay/notification/internal/type/users/provider/UsersProvider.java
@@ -14,7 +14,7 @@
 
 package com.liferay.notification.internal.type.users.provider;
 
-import com.liferay.notification.type.NotificationContext;
+import com.liferay.notification.context.NotificationContext;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.User;
